extract Midi connector logic to hook
