+++ src/components/App/style.sass
@@ -1,0 +1,2 @@
+.header
+  text-align: center
