+++ src/components/MidiConnector/index.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { useMidi } from '../../hooks/midi';
+
+const MidiConnector = () => {
+  const { handleSelectChange, midiInputId, midiInputs } = useMidi();
+
+  const inputOptions = () => {
+    for (const input of midiInputs) {
+      return <option value={input.id}>{input.name}</option>;
+    }
+  };
+
+  return (
+    <select value={midiInputId} onChange={handleSelectChange}>
+      {midiInputs.length ? (
+        inputOptions()
+      ) : (
+        <option value="noinput" disabled>
+          No MIDI devices
+        </option>
+      )}
+    </select>
+  );
+};
+
+export default MidiConnector;
