+++ src/components/Pads/style.sass
@@ -1,0 +1,26 @@
+.pads
+  display: grid
+  grid-template-columns: repeat(4, 1fr)
+  grid-auto-rows: 1fr
+  &::before
+    content: ''
+    width: 0
+    padding-bottom: 100%
+    grid-row: 1 / 1
+    grid-column: 1 / 1
+
+  & > *:first-child
+    grid-row: 1 / 1
+    grid-column: 1 / 1
+
+  @media (max-width: 37.5em)
+    grid-template-columns: repeat(3, 1fr)
+  @media (max-width: 25.875em)
+    grid-template-columns: repeat(2, minmax(var(--pad-min-size), 1fr))
+
+.pad
+  background-color: rgba(0, 0, 0, 0.1)
+  border: 1px white solid
+
+  &.lit
+    background-color: blue
