+++ src/components/Selector/style.sass
@@ -1,0 +1,7 @@
+.selector
+  font-size: .8rem
+  @media (max-width: 48em)
+    display: none
+
+.selector-sample
+  padding: .1rem
